Ispravke sitnijih bugova i izmene za navmeni.

+++ ProductionApp/front/app/src/main/java/com/example/pyxiskapri/dtos/response/PostOnMapResponse.kt
@@ -14,7 +14,7 @@ data class PostOnMapResponse(
     @SerializedName("dateCreated")
     var postDate: String,
     @SerializedName("numberOfLikes")
-    var likeCount: Int,
+    var likeCount: Double,
     @SerializedName("numberOfViews")
     var viewCount: Int,
     @SerializedName("isLiked")
